Route seates fix

+++ CarPool/CarPool/Controllers/RouteController.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using CarPool.Data;
 using CarPool.Models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -35,6 +37,7 @@ namespace CarPool.Controllers
         {
             ModelState.Clear(); // zamezi zobrazeni validačni hlašky po prvním spuštení
                                 //  return View("Clear");
+
             return View("FindRouteForm");
         }
 
@@ -129,23 +132,48 @@ namespace CarPool.Controllers
         [Authorize]
         [HttpPost]
         [Route("/Route/JoinRoute")]
-        public void JoinRoute(int id)
+        public string JoinRoute(int id)
         {
-            _db.routeUsers.Add(new RouteUser { RouteId = id, UserId = userManager.GetUserId(User) });
-            _db.SaveChanges();
+            Route routeFromDB = new Route();
+            routeFromDB = _db.routes.Where(r => r.id == id).FirstOrDefault();
+            if (routeFromDB != null)
+            {
+                //try if exists
+                RouteUser exists = _db.routeUsers.Where(ru => ru.RouteId == id && ru.UserId == userManager.GetUserId(User)).FirstOrDefault();
+                if (routeFromDB.seats > 0 && exists==null)
+                {
+                    routeFromDB.seats--;
+                    _db.SaveChanges();
+                    _db.routeUsers.Add(new RouteUser { RouteId = id, UserId = userManager.GetUserId(User) });
+                    _db.SaveChanges();
+                }
+                else
+                {
+                    TempData["msg-error"] = "Tato jízda je již plná";
+                }
+                
+            }
+            return "OK";
         }
 
         [Authorize]
         [HttpPost]
         [Route("/Route/LeaveRoute")]
-        public void LeaveRoute(int id)
+        public string LeaveRoute(int id)
         {
-            ViewData["Message"] = "Your application description page.";
-            Route routeFromDb = new Route();
+            var routeTodelete = _db.routeUsers.Where(ru => ru.RouteId == id && ru.UserId == userManager.GetUserId(User)).FirstOrDefault();
+            if (routeTodelete != null)
+            {
 
-                var routeTodelete = _db.routeUsers.Where(ru => ru.RouteId == id && ru.UserId == userManager.GetUserId(User)).FirstOrDefault();
+                Route r = _db.routes.Where(ro => ro.id == id).FirstOrDefault();
+                r.seats++;
+
                 _db.routeUsers.Remove(routeTodelete);
                 _db.SaveChanges();
+            }
+            else
+                TempData["msg-error"] = "Došlo k chybě při odhlašování";
+            return "OK";
         }
 
         [Authorize]
@@ -153,22 +181,22 @@ namespace CarPool.Controllers
         public IActionResult ShowLoggedRoutes()
         {
             List<Route> myRoutes = new List<Route>();
-       
-
-       
-                var query = from RouteUser in _db.routeUsers
-                            join route in _db.routes on RouteUser.RouteId equals route.id
-                            where RouteUser.UserId == userManager.GetUserId(User) && route.id == RouteUser.RouteId
-                            select route;
 
 
-                foreach (Route routeFromQuery in query)
-                {
-                    routeFromQuery.connected = true;
-                    myRoutes.Add(routeFromQuery);
-                }
 
-            
+            var query = from RouteUser in _db.routeUsers
+                        join route in _db.routes on RouteUser.RouteId equals route.id
+                        where RouteUser.UserId == userManager.GetUserId(User) && route.id == RouteUser.RouteId
+                        select route;
+
+
+            foreach (Route routeFromQuery in query)
+            {
+                routeFromQuery.connected = true;
+                myRoutes.Add(routeFromQuery);
+            }
+
+
 
             ViewData["Routes"] = myRoutes;
             ViewData["Title"] = "Přihlášené jízdy";
@@ -187,15 +215,15 @@ namespace CarPool.Controllers
         {
             List<Route> createtByUser = new List<Route>();
 
-       
-                var query = _db.routes.Where(r => r.CreatedBy.Id == userManager.GetUserId(User));
 
-                foreach (Route route in query)
-                {
-                    route.connected = null;
-                    createtByUser.Add(route);
-                }
-         
+            var query = _db.routes.Where(r => r.CreatedBy.Id == userManager.GetUserId(User));
+
+            foreach (Route route in query)
+            {
+                route.connected = null;
+                createtByUser.Add(route);
+            }
+
             ViewData["Routes"] = createtByUser;
             ViewData["Title"] = "Vytvořené jízdy";
 
@@ -213,24 +241,28 @@ namespace CarPool.Controllers
         {
 
 
-                Route ro = new Route();
-                ro = _db.routes.Where(r => r.id == id).FirstOrDefault();
+            Route ro = new Route();
+            ro = _db.routes.Where(r => r.id == id && r.CreatedBy.Id == userManager.GetUserId(User)).FirstOrDefault();
 
-                if (ro.CreatedBy.Id.Equals(userManager.GetUserId(User)))
+            if (ro != null)
+            {
+
+                var routesToDelete = _db.routeUsers.Where(r => r.RouteId == id);
+                foreach (RouteUser ru in routesToDelete)
                 {
-
-                    var routesToDelete = _db.routeUsers.Where(r => r.RouteId == id);
-                    foreach (RouteUser ru in routesToDelete)
-                    {
-                        _db.Remove(ru);
-                    }
-
-
-                    ro = _db.routes.Where(r => r.id == id && r.CreatedBy.Id == userManager.GetUserId(User)).FirstOrDefault();
-                    _db.routes.Remove(ro);
-                    _db.SaveChanges();
+                    _db.Remove(ru);
                 }
-          
+
+
+                ro = _db.routes.Where(r => r.id == id && r.CreatedBy.Id == userManager.GetUserId(User)).FirstOrDefault();
+                _db.routes.Remove(ro);
+                _db.SaveChanges();
+            }
+            else
+            {
+                TempData["msg-error"] = "Nastala chyba při mazaní jízdy";
+            }
+
 
 
         }
